fix price to display always in #.00 notation price is also visible in popup when ordering

+++ backend/WebApi/Lunchorder.Domain/Entities/DocumentDb/MenuRule.cs
@@ -26,7 +26,7 @@ namespace Lunchorder.Domain.Entities.DocumentDb
         /// <summary>
         /// The positive or negative price difference
         /// </summary>
-        public int PriceDelta { get; set; }
+        public double PriceDelta { get; set; }
 
         /// <summary>
         /// Field that represents the enabled state
